+++ tegral-prismakt/tegral-prismakt-generator-tests/mysql-types/src/test/kotlin/guru/zoroark/tegral/prismakt/generator/tests/mysqltypes/MysqlTypesTest.kt
@@ -52,7 +52,6 @@ class MysqlTypesTest {
             driver = "com.mysql.cj.jdbc.Driver",
             user = "root",
             password = "root"
-
         )
     }